[Embedded] Split dynamic casting machinery into a separate file
